Function serverdyna::handle_client

pub(crate) async fn handle_client(
    socket: TcpStream,
    user: String,
    ip: String,
    server: String
)

Gère la connexion d’un client, enregistre les détails de la connexion et envoie une réponse.

Cette fonction est exécutée de manière asynchrone pour chaque client connecté.

Arguments

socket - Un objet TcpStream représentant la connexion du client.

user - Une String représentant le nom de l’utilisateur.

ip - Une String représentant l’adresse IP du client.

server - Une String représentant l’adresse du serveur.

Examples

tokio::spawn(handle_client(socket, "user1".to_string(), "127.0.0.1".to_string(), "127.0.0.1:8080".to_string()));

Panics

Cette fonction ne devrait pas paniquer dans des conditions normales d’utilisation.

Errors

Cette fonction enregistre les erreurs dans la sortie standard d’erreurs (stderr) lorsqu’elles se produisent.